adjusting

+++ KonySim/KonySim/FightCalculator.cs
@@ -13,7 +13,7 @@ namespace KonySim
         public static void MissionFight(Db.Connection con, List<Db.Soldier> soldiers, Db.Mission mission)
         {
             float soldiersPower = 0;
-            float missionPower = mission.CivilianCount * (mission.DefenseMultiplier + r.Next(4)) * (1/99) +
+            float missionPower = (float)(mission.CivilianCount * (mission.DefenseMultiplier + r.Next(4))) * (1/99) +
                 mission.AnimalCount * (mission.DefenseMultiplier + r.Next(4)) * 2;
             
             foreach (Db.Soldier s in soldiers)
@@ -25,13 +25,14 @@ namespace KonySim
                     weaponPower += w.Damage;
                 }
 
-                soldiersPower += (0/*replace 0 with drug bool*/ + 1 + (s.Lvl / 5)) * weaponPower;
+                soldiersPower += (0/*replace 0 with drug bool*/ + 1 + ((float)s.Lvl / 5)) * weaponPower;
             }
 
             ///*
             /// Win or loss of battle is administrated below.
             /// */
 
+            //Winning battle.
             if (soldiersPower >= missionPower)
             {
                 //Child stuff.
@@ -57,6 +58,7 @@ namespace KonySim
                 mission.AnimalCount = 0;
                 mission.Completed = true;
             }
+            //Losing battle.
             else
             {
                 float powerDifference = 1 / (1 + ((missionPower - soldiersPower) / soldiersPower));
@@ -65,7 +67,7 @@ namespace KonySim
 
                 foreach (Db.Soldier s in soldiers)
                 {
-                    s.Exp += (int)totalExp / soldiers.Count;
+                    s.Exp += (int)((float)totalExp / (float)soldiers.Count);
                     s.Lvl = (int)(Math.Pow((double)s.Exp, 0.8) / 100);
                     s.Health -= (int)Math.Pow((double)r.Next(25, 201), Math.Log(100, 200));
